pyproject.toml
[project]
name = "okenaba"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.39.0",
    "google-analytics-data>=0.20.0",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
    "python-fasthtml>=0.12.0",
    "supabase>=2.0.0",
    "psycopg2-binary>=2.9.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
